Use a dependency group instead of optional dependencies
Dependency groups are relatively new. They are a better way of providing
tools for a project that are only for development and testing purposes.
Optional dependencies are those that should ship with the final project to
provide extra functionality. This makes dependency groups a better fit for
the tools that were listed in optional dependecies.

https://peps.python.org/pep-0735/
https://github.com/pypa/pip/issues/12963
https://github.com/pypa/pip/pull/13065