Restore the main license to the repo root
The only magic parsing is of a top level LICENSE file and text
inside the README. To make it clear what is the main project license,
it should left at the repo root. Third part license files go in licenses/.
All are included in the container image for documentation purposes.