Transparently vendor external packages (#59)
* Add transparent vendoring

Insert paths to sys.path for every module inside _vendor. Call
this early in order to setup sys.path before any other imports happen.

* Update project license

This code came from Ansible, which is GPL 3.

* Ignore vendored packages

* Add tests for vendor

* Update doc string

* Change how modules are imported so they are releaded reliabliy

* Restore the main license to the repo root

The only magic parsing is of a top level LICENSE file and text
inside the README. To make it clear what is the main project license,
it should left at the repo root. Third part license files go in licenses/.
All are included in the container image for documentation purposes.