Add a helper function to test whether an executable exists on a remote system (#71)
* Add a helper function for getting the pull path of an executable on a remote host

This allows testing for whethere or not a command exists
before the tool tries to use it and report a more useful
failure to the LLM client.

* Use get_bit_path with a few tools as an example

* Create separate functions for local and remote

* Remove username as a parameter

Rather than a paremeter, this should come from config. The config value should
default to None so that the ssh_config settings can be used.

This may cause some problems with the ssh connection caching as it is now
by host.

* Move command lookup into execute_command

This may be a bit too helpful and frustrate the caller in the future, but
works well given the current state of the code.

* Search sbin paths by default

* Raise ToolError for all tools

* Update tests

* Provide a spec to avoid incorrect mocking and remove warning

* Log ssh username using value from the connection

This lets async ssh do the work of getting the username based on ssh_config
which is more in line with user expectations.

* Fixup tests

* Append to PATH more succinctly

* Separate ssh tests into separate modules

* Fixup more tests

More accurate and less verbose patching

* Move command lookup into the local and remote execution functions

This gets rid of another ‘if host’ branch in the code. Now the local and remote
exeucution code takes care of looking up the command in the way that makes sense
for that context.

* Correct typeo in docs

* Pass in unpacked args

* Change execption raised when file is not found

* Let exceptions bubble up to tool call

The framework catches any raised execption in a tool and raises a ToolError.
There is no need to explicitly raise a ToolError since any exception in a tool
ends up being raised as a ToolError.

* Improve error test cases for log tools

* Let the exception bubble up

* Consolidate tests

* Simplify key discovery tests

* Make mocks simpler and add additional test case

* Add explicit tests for get_bin_path and get_remote_bin_path

* Do not catch exception raised by get_remote_bin_path

Let it bubble all the way up.

* Bring some sanity to the connection manager tests

* Fix typo