Use get_command instead of accessing global dictionary
All tools were accessing the global dictionary of `COMMANDS` directly,
and some of them had condition handling in case the specified command
was not in the proper state, but many of them didn't. This patch moves
the conditional handling to be inside `get_command` and refactor the
tools to use it.